requirements.txt
dask
pandas
numpy
hydra-core
nltk
emoji
textblob
pyarrow